Update patch set 1
Patch Set 1:

I think we need to move this to a subdirectory rather than in the root of the repo ... e.g. CCSDK-A1-Controller/  ???

Although this a large initial seed code contribution most of this contribution came from elsewhere, (SDNC repos?). Please add a mention of this in the Readme, along with a brief sentence or two about the new parts (northbound/nonrt-ric-api/* etc)

Is there a need for all of the OAM and GenricResourceAPI stuff? 

Thanks

Patch-set: 1
Label: Verified=0