Update patch set 2
Patch Set 2:

There is a need for managing the functionality in the OSC NonRT RIC. This cannot have any dependency to the OSC NearRT RIC (other NearRT RIC implementations can exists in a multi vendor environment). 
The OSC NearRT RIC Dashboard is managing the OSC NearRT RIC (exclusively). Hence there is a need for two independent dashboards. This was a quick start. The non used code will be removed in later commits. 
The functionality of the two dashboards are already completely different and in the future I expect that it will be even more so. 

I think we can elaborate if it is possible to extract the common parts into a common repo to avoid maintaining the same code twice. I am not clear about how much code that actually will be the same though.

Patch-set: 2