Update patch set 2
Patch Set 2:

I hear you saying, this is the starting point for a totally new web application.  Whatever you do, please choose a name other than "dashboard".  I don't feel that name is very descriptive for the RIC implementation but for the time being, I'm stuck with it.  You get to use a better one.

Patch-set: 2