Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> What is this change?
> 
> There should certainly not be a dashboard/INFO.yaml
> 
> Please revert.

The commit removes the dashboard/INFO.yaml, so the comment is pretty strange.  
This code is far from production status and will likely be moved to an own repo in a later stage. 
If you want to discuss the scope of the project, please contact John Keeney.

Patch-set: 1