Update patch set 2
Patch Set 2: Code-Review+1

OK for now, but I think we should not write a lot of code and tests that is synchronous when we will have to change that later to asynch (double work).

The startup will be slow if there are many RICs and everything is run in series.

Patch-set: 2
Reviewer: Gerrit User 1000262 <1000262@84e4e190-e623-4c37-9eda-58b1937c1b53>
Label: Code-Review=+1