Update patch set 1
Patch Set 1:

Hi Cris, 

I see this way of selecting RIC as a temporary solution (by a single select drop-down). This might scale up to <100 RICs in the network.
But when there are thousands, this will not do. There is some more thinking needed in the SMO layer. The natural way is maybe (usually) to group them hierarchically in some way. 

When it comes to configuring of policies, that must be done by the SMO layer (which also is according to the architecture).
We are in the process of developing a 'policy agent' (working name). 
It will monitor all R-app, RICs and Policies. 
If there are any policy inconsistencies between the SMO view and the RIC, policies will be deleted from the RIC. 
When the service (R-APP) that "owns" a policy is 'dead', all its policies will be automatically deleted. 
When the agent is restarted, all policies will be deleted from all RICs.

This 'dashboard' will work  towards the NBI of this 'agent' (in the same way as any R-App will do). 

So, having a way to configure policies directly towards the RIC might be good to have when experimenting, but I do not think it will work in a real system. 
Any policy not created via this SMO 'policy agent' will be deleted.

This was my view. 

I would be great if we could meet and share our views! 

BR Patrik Buhr

Patch-set: 1
Reviewer: Gerrit User 1000262 <1000262@84e4e190-e623-4c37-9eda-58b1937c1b53>
Label: Code-Review=0