Update patch set 2
Patch Set 2:

Looking at the logs, the problem seems to be with the naming of jar file. The merge job builds jar with the name "policy-agent-1.0.0-SNAPSHOT.jar" and this name is actually hard-coded in the dockerfile. Hence, this job succeeds. However, the staging job builds jar with the name "policy-agent-1.0.0.jar" (for some reason, it removes SNAPSHOT at the end) and this leads to failure. I tried to remove the hard-coded name of jar file in the commit below, and hopefully that should fix the problem: https://gerrit.o-ran-sc.org/r/#/c/nonrtric/+/2167/

Patch-set: 2
Reviewer: Gerrit User 1000128 <1000128@84e4e190-e623-4c37-9eda-58b1937c1b53>
Label: Code-Review=0