Recovery handling
Supervision of NearRT RIC policy types and instances.
When there is a difference between the view in the NonRT RIC
and the NearRT RIC, all policies will be deleted and the
types will be refreshed.

Introduced that policies will be created and deleted in the NonRT RIC
when created/deleted from the agent NBI.

Added Mock mode, run by command:
mvn -Dtest=MockPolicyAgent test

Issue-ID: NONRTRIC-84
Change-Id: Ica621c990f352cd69efa0dca51451d5f3c755b68
Signed-off-by: PatrikBuhr <patrik.buhr@est.tech>