Update patch set 1
Patch Set 1:

(1 comment)

When I run the ApplicationTest with coverage from inside Eclipse it breaks and there are java.lang.StackOverflowError errors and the test is left hanging. I then tried to add a System.out.println in the Lock::processQueuedEntries method, and got the same result. So the solution seems to be a bit brittle when it comes to timing.

Patch-set: 1
Reviewer: Gerrit User 1000201 <1000201@84e4e190-e623-4c37-9eda-58b1937c1b53>