Update patch set 1
Patch Set 1:

(1 comment)

> Patch Set 1:
> 
> (1 comment)
> 
> When I run the ApplicationTest with coverage from inside Eclipse it breaks and there are java.lang.StackOverflowError errors and the test is left hanging. I then tried to add a System.out.println in the Lock::processQueuedEntries method, and got the same result. So the solution seems to be a bit brittle when it comes to timing.

It has to to do with the used stack size (not timing). 
The concurrency test is a bit exhausting, it runs 100 threads where each creates/deletes 100 policies and runs 10 consistency checks. 
I will take a look if the stack usage can be decreased (maybe processQueuedEntries must be run in a separate thread).

Patch-set: 1