Concurrency improvements
The test for concurrency is improved so it involves RIC
synchronizations.

The principles for synchronization is simplified so that
classes in repository always returns copies of collections.

RIC syncronization is taking an exclusive lock during syncronization,
which leads to that a client will not get an HTTP error when accessing
a RIC that is synched. Instead, the client will be kept waiting until
the synch is completed.

Change-Id: I67568e8ef63b4b559a341ed8136e41119c9b7e6b
Issue-ID: NONRTRIC-164
Signed-off-by: PatrikBuhr <patrik.buhr@est.tech>