Loading of configuration
When an invalid configuartion was entered (from file of from consul)
the periodic polling for updates was stopped. This is fixed.

Improved error messages when an invalid configiration was intered.
Previuosly it often was just a NullPointerException when a mandatory
element was not found. Now, the missing element will be logged.

Change-Id: Id28dd268429c0174a84f490e6fb48ef5614c0966
Issue-ID: NONRTRIC-204
Signed-off-by: PatrikBuhr <patrik.buhr@est.tech>