Update patch set 10
Patch Set 10:

(2 comments)

In the last successful run of the test I cannot see that FTC10 is run at all, in any of the logs. In the last failing one I can see that it has been run. Has FTC10 really been run in the last test? And if so, why can't I see it in the logs?

Patch-set: 10
Reviewer: Gerrit User 1000201 <1000201@84e4e190-e623-4c37-9eda-58b1937c1b53>