Update patch set 10
Patch Set 10:

> Patch Set 10:
> 
> (2 comments)
> 
> In the last successful run of the test I cannot see that FTC10 is run at all, in any of the logs. In the last failing one I can see that it has been run. Has FTC10 really been run in the last test? And if so, why can't I see it in the logs?

We have two jobs for each commit, it is run in job "maven docker verify":
https://jenkins.o-ran-sc.org/job/nonrtric-maven-docker-verify-master-mvn36-openjdk11/681/consoleFull
you can see the log between key words "run_integration" and "run_integration END"

as for the other job "maven verify", this step is skipped because the VM in LF for this job does not support docker.
https://jenkins.o-ran-sc.org/job/nonrtric-policy-agent-maven-verify-master-mvn36-openjdk11/22/console

I made a commit to LF jenkins repo to skip this step for this job.
https://gerrit.o-ran-sc.org/r/c/ci-management/+/4024

Patch-set: 10