Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Isn't it better that we use OpenAPI and define the API in a json or yaml file and generate the code for the controllers?

I think that is not obvious. Personally, I think it is very comfortable to work with the ECS and generate the doc. The tags used can be constants on code, but would have to be repeated in yaml. 

If we are going to implement any kind of authorization, I have a feeling we should be as close to the REST as possible. Probably there will be some kind of ticket in REST header that we need to get a hold of (and I am not sure how that is possible).

Patch-set: 1
Reviewer: Gerrit User 1000262 <1000262@84e4e190-e623-4c37-9eda-58b1937c1b53>